+++ pref/lib/src/dialog.dart
@@ -3,15 +3,12 @@
 // Use of this source code is governed by a MIT license that can be
 // found in the LICENSE file.
 
-import 'dart:async';
-
 import 'package:flutter/material.dart';
 
-import 'service/base.dart';
-import 'service/cache.dart';
 import 'service/pref_service.dart';
+import 'service_cache.dart';
 
-class PrefDialog extends StatefulWidget {
+class PrefDialog extends PrefCache {
   const PrefDialog({
     Key key,
     @required this.children,
@@ -21,49 +18,26 @@ class PrefDialog extends StatefulWidget {
     this.dismissOnChange = false,
     this.cancel,
   })  : assert(children != null),
-        onlySaveOnSubmit = onlySaveOnSubmit ?? submit != null,
         assert(dismissOnChange != null),
-        super(key: key);
+        super(key: key, cache: onlySaveOnSubmit ?? submit != null);
 
   final Widget title;
   final List<Widget> children;
   final Widget submit;
   final Widget cancel;
 
-  final bool onlySaveOnSubmit;
   final bool dismissOnChange;
 
   @override
   PrefDialogState createState() => PrefDialogState();
 }
 
-class PrefDialogState extends State<PrefDialog> {
-  Future<BasePrefService> _cache;
-
+class PrefDialogState extends PrefCacheState<PrefDialog> {
   @override
-  void didChangeDependencies() {
-    if (widget.onlySaveOnSubmit) {
-      _cache ??= _createCache();
-    }
-    super.didChangeDependencies();
-  }
-
-  BasePrefService get _parent {
-    final parent = PrefService.of(context);
-
-    // Check if we already have a BasePrefService
-    if (parent == null) {
-      throw FlutterError(
-          'No PrefService widget found in the tree. Unable to load settings');
-    }
-
-    return parent;
-  }
-
-  Widget _buildDialog(BuildContext context, BasePrefService parent) {
+  Widget buildChild(BuildContext context) {
     final actions = <Widget>[];
 
-    if (widget.cancel != null && widget.onlySaveOnSubmit) {
+    if (widget.cancel != null && widget.cache) {
       actions.add(
         FlatButton(
           child: widget.cancel,
@@ -78,10 +52,8 @@ class PrefDialogState extends State<PrefDialog> {
       actions.add(
         FlatButton(
           child: widget.submit,
-          onPressed: () {
-            if (widget.onlySaveOnSubmit) {
-              parent.apply(PrefService.of(context));
-            }
+          onPressed: () async {
+            await apply();
             Navigator.of(context).pop();
           },
         ),
@@ -91,11 +63,9 @@ class PrefDialogState extends State<PrefDialog> {
     if (widget.dismissOnChange) {
       Function f;
 
-      f = () {
+      f = () async {
         PrefService.of(context).removeListener(f);
-        if (widget.onlySaveOnSubmit) {
-          parent.apply(PrefService.of(context));
-        }
+        await apply();
         Navigator.of(context).pop();
       };
 
@@ -111,45 +81,5 @@ class PrefDialogState extends State<PrefDialog> {
       ),
       actions: actions,
     );
-  }
-
-  Future<BasePrefService> _createCache() async {
-    final service = PrefServiceCache();
-    await service.apply(_parent);
-    return service;
-  }
-
-  Widget _buildService(BuildContext context, BasePrefService parent) {
-    if (widget.onlySaveOnSubmit) {
-      return FutureBuilder(
-        future: _cache,
-        builder: (BuildContext context, snapshot) {
-          if (!snapshot.hasData) {
-            return const SizedBox();
-          }
-
-          return PrefService(
-            service: snapshot.data,
-            child: Builder(
-              builder: (BuildContext context) => _buildDialog(context, parent),
-            ),
-          );
-        },
-      );
-    }
-
-    return _buildDialog(context, parent);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    // Check if we already have a BasePrefService
-    final service = PrefService.of(context);
-    if (service == null) {
-      throw FlutterError(
-          'No PrefService widget found in the tree. Unable to load settings');
-    }
-
-    return _buildService(context, service);
   }
 }